+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+ - `ctags` like handling for `-f <tagfile>` command line args param
+ - Add support for reading command arguments from single file or directory with `.ctags` files
+ - Add basic tagfile validation for existing tagfile in `append` mode
+
+### Removed
+ - Recursive search for tag file in current directory and it's parent folders
+
 ## [0.3.2] 2025-10-28
 
 ### Added